+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.0.2
+##  20-01-2023
+
+1. [](#new)
+    * Added auto-clean job to limit local blacklist size
+2. [](#improved)
+    * Added more filter patterns for common exploits
+
 # v1.0.1
 ##  19-01-2023